Implementando classe universitária

+++ Desafios/BancoPOO/Poupanca.java
@@ -3,7 +3,7 @@ package Desafios.BancoPOO;
 //Poupança só pode sacar 200 reais
 public class Poupanca extends Conta{
     private
-        int limiteSaque = 200;
+        Double limiteSaque = 200D;
 
     public Poupanca(Cliente portador, String codigo, Double saldo){
         super(portador, codigo, saldo);
